implement some features to manage token count
Add:
- Count and truncate by tokens amount (#40)
- Sort tags by token count

Change:
- Move "use raw clip token..." setting to "Settings" tab
- Show tag frequency, length or token count depending on the "Sort by"
